requirements.txt
# ══════════════════════════════════════════════════════════════════
# Core — LangGraph + LangChain
# ══════════════════════════════════════════════════════════════════
langgraph==0.2.28
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-community==0.3.5

# ══════════════════════════════════════════════════════════════════
# Search tools
# ══════════════════════════════════════════════════════════════════
tavily-python==0.5.0          # Best search for research agents (recommended)
duckduckgo-search==6.3.3      # Free fallback — no API key required

# ══════════════════════════════════════════════════════════════════
# Web scraping
# ══════════════════════════════════════════════════════════════════
requests==2.32.3

# ══════════════════════════════════════════════════════════════════
# API server
# ══════════════════════════════════════════════════════════════════
fastapi==0.115.0
uvicorn[standard]==0.31.0     # [standard] includes extras for production
pydantic==2.9.2

# ══════════════════════════════════════════════════════════════════
# Environment
# ══════════════════════════════════════════════════════════════════
python-dotenv==1.0.1
